add Position struct
